[kernel] Build all possible drivers whether linked into kernel or not
