+++ 8018x.config
@@ -89,7 +89,6 @@ CONFIG_EXEC_MMODEL=y
 # CONFIG_BLK_DEV_FD is not set
 # CONFIG_BLK_DEV_BHD is not set
 # CONFIG_BLK_DEV_FD is not set
-# CONFIG_BLK_DEV_HD is not set
 # CONFIG_TRACK_CACHE is not set
 # CONFIG_IDE_PROBE is not set
 # CONFIG_BLK_DEV_RAM is not set
