+++ blink16.config
@@ -68,7 +68,6 @@ CONFIG_BLK_DEV_BFD=y
 CONFIG_BLK_DEV_BHD=y
 CONFIG_IDE_PROBE=y
 # CONFIG_BLK_DEV_FD is not set
-# CONFIG_BLK_DEV_HD is not set
 # CONFIG_BLK_DEV_RAM is not set
 CONFIG_RAMDISK_SEGMENT=0xD000
 CONFIG_RAMDISK_SECTORS=128
