+++ elks/arch/i86/drivers/block/unused/directhd.c
