+++ elks/arch/i86/drivers/char/init.c
@@ -8,10 +8,6 @@ void INITPROC chr_dev_init(void)
     lp_init();
 #endif
 
-#ifdef CONFIG_CHAR_DEV_CGATEXT
-    cgatext_init();
-#endif
-
 #ifdef CONFIG_CHAR_DEV_MEM
     mem_dev_init();
 #endif
