+++ emu86-disk.config
@@ -65,7 +65,6 @@ CONFIG_TRACK_CACHE=y
 CONFIG_BLK_DEV_BHD=y
 CONFIG_IDE_PROBE=y
 # CONFIG_BLK_DEV_FD is not set
-# CONFIG_BLK_DEV_HD is not set
 CONFIG_BLK_DEV_RAM=y
 CONFIG_RAMDISK_SEGMENT=0
 CONFIG_RAMDISK_SECTORS=128
