+++ emu86-rom-full.config
@@ -87,7 +87,6 @@ CONFIG_BLK_DEV_BFD=y
 CONFIG_BLK_DEV_BHD=y
 # CONFIG_IDE_PROBE is not set
 # CONFIG_BLK_DEV_FD is not set
-# CONFIG_BLK_DEV_HD is not set
 # CONFIG_BLK_DEV_RAM is not set
 CONFIG_BLK_DEV_CHAR=y
