+++ emu86-rom.config
@@ -83,7 +83,6 @@ CONFIG_ROMFS_BASE=8000
 # CONFIG_BLK_DEV_FD is not set
 # CONFIG_BLK_DEV_BHD is not set
 # CONFIG_BLK_DEV_FD is not set
-# CONFIG_BLK_DEV_HD is not set
 # CONFIG_BLK_DEV_RAM is not set
 CONFIG_BLK_DEV_CHAR=y
