+++ image/Make.devices
@@ -189,12 +189,6 @@ endif
 	$(MKDEV) /dev/le0 c 9 4
 
 ##############################################################################
-# CGATEXT
-ifdef CONFIG_CHAR_DEV_CGATEXT
-	$(MKDEV) /dev/cgatext	c 10 0
-endif
-
-##############################################################################
 # ROM / Flash memory card
 
 ifdef CONFIG_ROMCODE
