+++ solo86-rom-full.config
@@ -84,7 +84,6 @@ CONFIG_FS_NR_EXT_BUFFERS=32
 # CONFIG_BLK_DEV_BHD is not set
 # CONFIG_IDE_PROBE is not set
 # CONFIG_BLK_DEV_FD is not set
-# CONFIG_BLK_DEV_HD is not set
 # CONFIG_BLK_DEV_RAM is not set
 CONFIG_BLK_DEV_ATA_CF=y
 CONFIG_BLK_DEV_CHAR=y
